+++ tests/test_signing.py
@@ -5,8 +5,6 @@ from __future__ import absolute_import, print_function
 import hashlib
 import unittest
 
-from six.moves.urllib.parse import parse_qs
-
 from laterpay import signing
 
 
@@ -78,24 +76,6 @@ class TestSigningHelper(unittest.TestCase):
         self.assertEqual(
             mac,
             '346f3d53ad762f3ed3fb7f2427dec2bbfaf0338bb7f91f0460aff15c',
-        )
-
-    def test_sign_and_encode(self):
-        params = {
-            u'parĄm1': u'valuĘ',
-            'param2': ['value2', 'value3'],
-            'ts': '1330088810',
-        }
-        url = u'https://endpoint.com/api'
-
-        secret = u'secret'  # unicode is what we usually get from api/db..
-
-        signed_and_encoded = signing.sign_and_encode(secret, params, url)
-
-        self.assertEqual(
-            signed_and_encoded,
-            'param2=value2&param2=value3&par%C4%84m1=valu%C4%98&'
-            'ts=1330088810&hmac=01c928dcdbbf4ba467969ec9607bfdec0563524d93e06df7d8d3c80d'
         )
 
     def test_verify_str_signature(self):
@@ -171,71 +151,6 @@ class TestSigningHelper(unittest.TestCase):
         false_params = params.copy()
         false_params['ts'] = '1234567890'
         self.assertFalse(signing.verify(false_params['hmac'], secret, false_params, url, method))
-
-    def test_signing_with_item(self):
-        secret = '401e9a684fcc49578c1f23176a730abc'
-        base_url = 'http://local.laterpaytest.net:8005/dialog/mmss/buy'
-        method = 'GET'
-
-        # creating this data with ItemDefinition and copy.copy(item.data) doesn't work
-        # since it has a purchase_date based on now(), so the signature isn't the same..
-        data = {
-            'article_id': 154,
-            'cp': ['laternews'],
-            'jsevents': [1],
-            'pricing': ['EUR200'],
-            'purchase_date': [1398861228815],
-            'title': [u"VIDEO: Rwanda's genocide, 20 years on"],
-            'tref': ['4ebbf443-a12e-4ce9-89e4-999ba93ba1dc'],
-            'ts': ['1398861228'],
-            'url': ['http://local.laterpaytest.net:8003/mmss/154'],
-        }
-
-        params = signing.sign_and_encode(secret, data, base_url, method)
-        expected_string = (
-            'article_id=154&'
-            'cp=laternews&'
-            'jsevents=1&'
-            'pricing=EUR200&'
-            'purchase_date=1398861228815&'
-            'title=VIDEO%3A+Rwanda%27s+genocide%2C+20+years+on&'
-            'tref=4ebbf443-a12e-4ce9-89e4-999ba93ba1dc&'
-            'ts=1398861228&'
-            'url=http%3A%2F%2Flocal.laterpaytest.net%3A8003%2Fmmss%2F154&'
-            'hmac=d51564b41c2a8719fcdcfc6bad46109d3b6c6f78afea4020d5801a3c'
-        )
-
-        self.assertEqual(expected_string, params)
-
-        # expected signature based on params above
-        signature = 'd51564b41c2a8719fcdcfc6bad46109d3b6c6f78afea4020d5801a3c'
-
-        self.assertTrue(signing.verify(signature, secret, data, base_url, method))
-
-        # changing the price in the url
-        false_string = (
-            'article_id=154&'
-            'cp=laternews&'
-            'jsevents=1&'
-            'pricing=EUR150&'
-            'purchase_date=1398861228815&'
-            'title=VIDEO%3A+Rwanda%27s+genocide%2C+20+years+on&'
-            'tref=4ebbf443-a12e-4ce9-89e4-999ba93ba1dc&'
-            'ts=1398861228&'
-            'url=http%3A%2F%2Flocal.laterpaytest.net%3A8003%2Fmmss%2F154&'
-            'hmac=4d41f1adcb7c6bf6cf9c5eb15b179fdbec667d53f2749e2845c87315'
-        )
-        false_params = parse_qs(false_string)
-
-        self.assertFalse(signing.verify(signature, secret, false_params, base_url, method))
-
-        # changing the base_url
-        false_base_url = 'http://local.laterpaytest.net:8005/dialog/mmss/add'
-        self.assertFalse(signing.verify(signature, secret, data, false_base_url, method))
-
-        # changing http method
-        false_method = 'POST'
-        self.assertFalse(signing.verify(signature, secret, data, base_url, false_method))
 
     def test_normalise_param_structure(self):
         params = {